+++ src/utils/platform.ts
@@ -1,9 +1,5 @@
 import { Platform } from "obsidian";
 
-export function supportsExternalFolders(): boolean {
-  return false;
-}
-
 export function supportsSecondScreenWindow(): boolean {
   return Platform.isDesktopApp;
 }